Flatten even more.
